Move cypher connector logic
