+++ packages/graph-explorer/src/connector/openCypher/fetchNeighbors/oneHopTemplate.ts
