+++ packages/graph-explorer/src/connector/openCypher/fetchNeighborsCount/neighborsCountTemplate.ts
