+++ packages/graph-explorer/src/connector/openCypher/fetchSchema/edgeLabelsTemplate.ts
