+++ packages/graph-explorer/src/connector/openCypher/fetchSchema/edgesSchemaTemplate.ts
