+++ packages/graph-explorer/src/connector/openCypher/fetchSchema/index.test.ts
@@ -1,6 +1,6 @@
 import { vi } from "vitest";
 import { SchemaResponse } from "@/connector/useGEFetchTypes";
-import fetchSchema from "./fetchSchema";
+import fetchSchema from ".";
 import { ClientLoggerConnector } from "@/connector/LoggerConnector";
 
 describe("OpenCypher > fetchSchema", () => {
