+++ packages/graph-explorer/src/connector/openCypher/fetchSchema/vertexLabelsTemplate.ts
