+++ packages/graph-explorer/src/connector/openCypher/fetchSchema/verticesSchemaTemplate.ts
