+++ packages/graph-explorer/src/connector/openCypher/fetchVertexTypeCounts/vertexTypeCountTemplate.ts
