+++ packages/graph-explorer/src/connector/openCypher/keywordSearch/index.ts
@@ -6,7 +6,7 @@ import type {
 } from "@/connector/useGEFetchTypes";
 import isErrorResponse from "@/connector/utils/isErrorResponse";
 import mapApiVertex from "../mappers/mapApiVertex";
-import keywordSearchTemplate from "../templates/keywordSearchTemplate";
+import keywordSearchTemplate from "./keywordSearchTemplate";
 import type { OCVertex } from "../types";
 import { OpenCypherFetch } from "../types";
