+++ packages/graph-explorer/src/connector/openCypher/keywordSearch/keywordSearchTemplate.ts
